test: 테스트 리팩터링
