+++ space-d/src/test/java/com/dnd/spaced/core/account/application/AccountServiceTest.java
@@ -9,6 +9,7 @@ import com.dnd.spaced.core.account.application.dto.request.ChangeCareerInfoReque
 import com.dnd.spaced.core.account.application.dto.request.ChangeProfileInfoRequest;
 import com.dnd.spaced.core.account.application.dto.response.AccountResponse;
 import com.dnd.spaced.core.account.application.exception.ForbiddenAccountException;
+import com.dnd.spaced.core.account.application.helper.AccountServiceTestHelper;
 import com.dnd.spaced.core.account.domain.Account;
 import com.dnd.spaced.core.account.domain.enums.ProfileImageName;
 import com.dnd.spaced.core.account.domain.enums.RegistrationId;
@@ -17,7 +18,6 @@ import com.dnd.spaced.core.account.domain.enums.exception.InvalidCompanyExceptio
 import com.dnd.spaced.core.account.domain.enums.exception.InvalidExperienceException;
 import com.dnd.spaced.core.account.domain.enums.exception.InvalidJobGroupException;
 import com.dnd.spaced.core.account.domain.enums.exception.InvalidProfileImageNameException;
-import com.dnd.spaced.core.account.domain.repository.AccountRepository;
 import java.util.Arrays;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayNameGeneration;
@@ -37,27 +37,13 @@ import org.springframework.transaction.annotation.Transactional;
 @SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
 @SuppressWarnings("NonAsciiCharacters")
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
-class AccountServiceTest {
+class AccountServiceTest extends AccountServiceTestHelper {
 
     @Autowired
     AccountService accountService;
 
-    @Autowired
-    AccountRepository accountRepository;
-
     @Test
     void 지정한_회원을_탈퇴_처리한다() {
-        // given
-        Account account = Account.builder()
-                                 .registrationId(RegistrationId.KAKAO)
-                                 .socialIdentifier("12345")
-                                 .nickname("재빠른지구001")
-                                 .profileImage("earth.png")
-                                 .role(Role.ROLE_USER)
-                                 .build();
-
-        accountRepository.save(account);
-
         // when
         accountService.withdrawal(account.getId());
 
@@ -68,7 +54,7 @@ class AccountServiceTest {
     @Test
     void 없거나_이미_탈퇴한_회원의_ID라면_아니라면_탈퇴할_수_없다() {
         // when & then
-        assertThatThrownBy(() -> accountService.withdrawal(1L))
+        assertThatThrownBy(() -> accountService.withdrawal(-999L))
                 .isInstanceOf(ForbiddenAccountException.class)
                 .hasMessage("존재하지 않는 회원이거나 이미 탈퇴한 회원입니다.");
     }
@@ -76,23 +62,13 @@ class AccountServiceTest {
     @Test
     void 회원_경력_정보를_변경한다() {
         // given
-        Account account = Account.builder()
-                                 .registrationId(RegistrationId.KAKAO)
-                                 .socialIdentifier("12345")
-                                 .nickname("재빠른지구001")
-                                 .profileImage("earth.png")
-                                 .role(Role.ROLE_USER)
-                                 .build();
-
-        accountRepository.save(account);
-
-        // when
         ChangeCareerInfoRequest request = new ChangeCareerInfoRequest(
                 "개발자",
                 "비공개",
                 "1~2년 차"
         );
 
+        // when
         accountService.changeCareerInfo(account.getId(), request);
 
         // then
@@ -109,23 +85,13 @@ class AccountServiceTest {
     @NullAndEmptySource
     void 유효한_직군_이름이_아니라면_경력_정보를_변경할_수_없다(String invalidJobGroupName) {
         // given
-        Account account = Account.builder()
-                                 .registrationId(RegistrationId.KAKAO)
-                                 .socialIdentifier("12345")
-                                 .nickname("재빠른지구001")
-                                 .profileImage("earth.png")
-                                 .role(Role.ROLE_USER)
-                                 .build();
-
-        accountRepository.save(account);
-
-        // when & then
         ChangeCareerInfoRequest request = new ChangeCareerInfoRequest(
                 invalidJobGroupName,
                 "비공개",
                 "1~2년 차"
         );
 
+        // when & then
         assertThatThrownBy(() -> accountService.changeCareerInfo(account.getId(), request))
                 .isInstanceOf(InvalidJobGroupException.class)
                 .hasMessageContaining("잘못된 직군 이름");
@@ -135,24 +101,14 @@ class AccountServiceTest {
     @NullAndEmptySource
     void 유효한_회사명이_아니라면_경력_정보를_변경할_수_없다(String invalidCompanyName) {
         // given
-        Account account = Account.builder()
-                                 .registrationId(RegistrationId.KAKAO)
-                                 .socialIdentifier("12345")
-                                 .nickname("재빠른지구001")
-                                 .profileImage("earth.png")
-                                 .role(Role.ROLE_USER)
-                                 .build();
-
-        accountRepository.save(account);
-
-        // when & then
         ChangeCareerInfoRequest request = new ChangeCareerInfoRequest(
                 "개발자",
                 invalidCompanyName,
                 "1~2년 차"
         );
 
-        assertThatThrownBy(() -> accountService.changeCareerInfo(account.getId(),request))
+        // when & then
+        assertThatThrownBy(() -> accountService.changeCareerInfo(account.getId(), request))
                 .isInstanceOf(InvalidCompanyException.class)
                 .hasMessageContaining("잘못된 회사 이름");
     }
@@ -161,23 +117,13 @@ class AccountServiceTest {
     @NullAndEmptySource
     void 유효한_경력이_아니라면_경력_정보를_변경할_수_없다(String invalidExperienceName) {
         // given
-        Account account = Account.builder()
-                                 .registrationId(RegistrationId.KAKAO)
-                                 .socialIdentifier("12345")
-                                 .nickname("재빠른지구001")
-                                 .profileImage("earth.png")
-                                 .role(Role.ROLE_USER)
-                                 .build();
-
-        accountRepository.save(account);
-
-        // when & then
         ChangeCareerInfoRequest request = new ChangeCareerInfoRequest(
                 "개발자",
                 "비공개",
                 invalidExperienceName
         );
 
+        // when & then
         assertThatThrownBy(() -> accountService.changeCareerInfo(account.getId(), request))
                 .isInstanceOf(InvalidExperienceException.class)
                 .hasMessageContaining("잘못된 경력");
@@ -185,14 +131,15 @@ class AccountServiceTest {
 
     @Test
     void 없거나_탈퇴한_회원의_ID라면_경력_정보를_변경할_수_없다() {
-        // when & then
+        // given
         ChangeCareerInfoRequest request = new ChangeCareerInfoRequest(
                 "개발자",
                 "비공개",
                 "1~2년 차"
         );
 
-        assertThatThrownBy(() -> accountService.changeCareerInfo(1L, request))
+        // when & then
+        assertThatThrownBy(() -> accountService.changeCareerInfo(-999L, request))
                 .isInstanceOf(ForbiddenAccountException.class)
                 .hasMessage("존재하지 않는 회원이거나 이미 탈퇴한 회원입니다.");
     }
@@ -206,23 +153,12 @@ class AccountServiceTest {
     @MethodSource("changeProfileInfoTestWithProfileImageKoreanName")
     void 회원_프로필_정보를_변경한다(ProfileImageName profileImageName) {
         // given
-        Account account = Account.builder()
-                                 .registrationId(RegistrationId.KAKAO)
-                                 .socialIdentifier("12345")
-                                 .nickname("재빠른지구001")
-                                 .profileImage("earth.png")
-                                 .role(Role.ROLE_USER)
-                                 .build();
-
-        account.changeCareerInfo("개발자", "비공개", "1~2년 차");
-        accountRepository.save(account);
-
-        // when
         ChangeProfileInfoRequest request = new ChangeProfileInfoRequest(
                 "행복한지구001",
                 profileImageName.getKorean()
         );
 
+        // when
         accountService.changeProfileInfo(account.getId(), request);
 
         // then
@@ -238,22 +174,12 @@ class AccountServiceTest {
     @NullAndEmptySource
     void 프로필_이미지_경로가_비어_있으면_프로필_정보를_변경할_수_없다(String invalidProfileImageKoreanName) {
         // given
-        Account account = Account.builder()
-                                 .registrationId(RegistrationId.KAKAO)
-                                 .socialIdentifier("12345")
-                                 .nickname("재빠른지구001")
-                                 .profileImage("earth.png")
-                                 .role(Role.ROLE_USER)
-                                 .build();
-
-        accountRepository.save(account);
-
-        // when & then
         ChangeProfileInfoRequest request = new ChangeProfileInfoRequest(
                 "재빠른지구001",
                 invalidProfileImageKoreanName
         );
 
+        // when & then
         assertThatThrownBy(() -> accountService.changeProfileInfo(account.getId(), request))
                 .isInstanceOf(InvalidProfileImageNameException.class)
                 .hasMessageContaining("잘못된 프로필 이미지 이름");
@@ -261,13 +187,14 @@ class AccountServiceTest {
 
     @Test
     void 없거나_탈퇴한_회원의_ID라면_프로필_정보를_변경할_수_없다() {
-        // when & then
+        // given
         ChangeProfileInfoRequest request = new ChangeProfileInfoRequest(
                 "재빠른지구001",
                 "earth.png"
         );
 
-        assertThatThrownBy(() -> accountService.changeProfileInfo(1L, request))
+        // when & then
+        assertThatThrownBy(() -> accountService.changeProfileInfo(-999L, request))
                 .isInstanceOf(ForbiddenAccountException.class)
                 .hasMessage("존재하지 않는 회원이거나 이미 탈퇴한 회원입니다.");
     }
@@ -306,7 +233,7 @@ class AccountServiceTest {
     @Test
     void 없거나_탈퇴한_회원의_ID라면_회원_정보를_조회할_수_없다() {
         // when & then
-        assertThatThrownBy(() -> accountService.findAccountInfo(1L))
+        assertThatThrownBy(() -> accountService.findAccountInfo(-999L))
                 .isInstanceOf(ForbiddenAccountException.class)
                 .hasMessage("존재하지 않는 회원이거나 이미 탈퇴한 회원입니다.");
     }
